backend/pyproject.toml
[tool.poetry]
name = "chatgpt-from-scratch"
version = "0.1.0"
description = "Creating ChatGPT from scratch"
authors = ["stephenCarmody <stephen.m.carmody@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^3.2.0"
torch = "^2.5.1"
sagemaker = "^2.237.3"
numpy = ">=1.9.0,<2.0"
python-dotenv = "^1.0.0"
boto3 = "^1.34.0"
transformers = "^4.48.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
